♻️ 7GHJ80 context: decompose verify task validators
