+++ .agentplane/tasks/202605290005-7GHJ80/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 7GHJ80 task: Context verify-task validator decomposition [202605290005-7GHJ80]
